Поправил файловую структуру Выделил компоненты шапки приложения, локаций, списка предложений
