+++ src/components/app/app.tsx
@@ -1,13 +1,15 @@
-import MainPage from '../pages/main-page';
+import MainPage from '../../pages/main-page/main-page';
+//import FavoritesPage from '../../pages/favorite-page/favorites-page';
+//import LoginPage from '../../pages/login-page/login-page';
+//import OfferPage from '../../pages/offer-page/offer-page';
 
 type AppProps = {
   cardsCount: number;
-  placesFound: number;
 }
 
-function App({cardsCount, placesFound}: AppProps): JSX.Element {
+function App({cardsCount }: AppProps): JSX.Element {
   return (
-    <MainPage cardsCount={cardsCount} placesFound={placesFound} />
+    <MainPage cardsCount={cardsCount} />
   );
 }
